Section 2 Functionality

+++ Models/Invoice.cs
@@ -4,10 +4,10 @@
     {
         public string InvoiceId { get; private set; }
         public DateTime DateIssued { get; private set; }
-        public double TotalAmount { get; private set; }
+        public decimal TotalAmount { get; private set; }
         public Order OrderReference { get; private set; }
 
-        public Invoice(Order order, double totalAmount)
+        public Invoice(Order order, decimal totalAmount)
         {
             InvoiceId = Guid.NewGuid().ToString();
             DateIssued = DateTime.Now;
